[1.0.X] Added a test from Jamie Gennis to ensure #9848 doesn't reappear.
The bug itself was fixed at some point in the past months (there have
been a few improvements to update() recently). Fixed #9848. 

(Merge of r10528 from trunk.  Per comment in #9848 the fix was in r9967,
which was backported in r9968, thus it seems right to backport the new
test as well.)


git-svn-id: http://code.djangoproject.com/svn/django/branches/releases/1.0.X@10607 bcc190cf-cafb-0310-a4f2-bffc1f526a37