[soc2009/http-wsgi-improvements] Establish the priorities and fallbacks for HttpResponseSendFile methods.
Change the setting to settings.HTTPRESPONSE_SENDFILE_METHOD. If this is set to None, we use handler methods,
but otherwise the header gets set, and we do not send any content. If neither of these are available,
use the FileWrapper fallback in HttpResponseSendFile.

This passes the test suite, but is untested on mod_python. I am still trying to figure out how to
view the headers of a response with the Content-Disposition "attachment."

git-svn-id: http://code.djangoproject.com/svn/django/branches/soc2009/http-wsgi-improvements@11268 bcc190cf-cafb-0310-a4f2-bffc1f526a37