[1.1.X] Fixed #12470 - encoding of comma and semi-colon in cookies.
This is a backport of r12282 from trunk.

The original bug was about CookieStorage, which does not exist in 1.1.X,
but the fix involved the underlying cookie storage.

The change fixes other bugs with cookies for Internet Explorer and Safari,
hence it is backported.  It could, however, also cause backwards
incompatibilities with existing javascript that may parse cookie values that
contain commas or semi-colons, and, very rarely, with existing cookie values
manipulated server side.


git-svn-id: http://code.djangoproject.com/svn/django/branches/releases/1.1.X@12283 bcc190cf-cafb-0310-a4f2-bffc1f526a37